Optimize memory usage of counties logic
Refactored the logic within the yield function to directly update zip-to-county mappings as data is read.
Additionally, moved the 'counties' package to the 'internal' directory to restrict its accessibility outside the module.
Removed the redundant existence check for zip codes in the map,

Tested:
    - Ran the program with a valid tar file and observed output.
    - Checked error messages for scenarios with a missing tar file argument.
    - Checked error messages for scenarios with an invalid tar file.
